requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
prophet
xgboost
fastapi
uvicorn
python-multipart
requests
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
joblib